test_uri: eliminate failing test
t_parse_uri5 suggests it was intended to diagnose an invalid port, but has
never passed.  Split it into two tests: one that verifies the URI with a
valid port, and a second that verifies that changing the URI produces an
error.  Then change coap_split_uri so that ports that cannot be represented
in coap_uri_t::port produce an error.